refactor: spellCheckAction 함수 리팩토링 (#267)
